refactor: delete onboarding folder
