+++ lib/presentation/onboarding/preparation_order/cubit/preparation_order_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:on_time_front/presentation/onboarding/cubit/onboarding/onboarding_cubit.dart';
+import 'package:on_time_front/presentation/onboarding/cubit/onboarding_cubit.dart';
 
 part 'preparation_order_state.dart';
