+++ lib/presentation/onboarding/screens/onboarding_screen.dart
@@ -9,7 +9,7 @@ import 'package:on_time_front/presentation/onboarding/preparation_name_select/sc
 import 'package:on_time_front/presentation/onboarding/preparation_time/cubit/preparation_time_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_time/screens/preparation_time_input_list.dart';
 import 'package:on_time_front/presentation/onboarding/components/schedule_spare_time_input.dart';
-import 'package:on_time_front/presentation/onboarding/cubit/onboarding/onboarding_cubit.dart';
+import 'package:on_time_front/presentation/onboarding/cubit/onboarding_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_order/cubit/preparation_order_cubit.dart';
 
 class OnboardingScreen extends StatelessWidget {
